# specs/002-backoffice-catalog-sales/openapi.yaml
openapi: 3.0.3
info:
  title: Pede Aqui — Backoffice Catalog, Sales & Reports (Spec 002)
  version: 0.1.0
  description: >
    Contract fragment for Spec 002. Merge into the backend OpenAPI at /api/v1.
    All endpoints require Bearer JWT; tenant resolved from JWT tenant_id or X-Tenant-Id.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /catalog/products/{productId}:
    patch:
      tags: [catalog]
      summary: Edit product attributes (partial). VENDOR_ADMIN own vendor; OPS/ADMIN tenant-wide.
      parameters: [{ $ref: '#/components/parameters/ProductId' }, { $ref: '#/components/parameters/TenantHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductUpdateRequest' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/ProductResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /catalog/products/{productId}/price:
    patch:
      tags: [catalog]
      summary: Edit the single-SKU price. May apply immediately or queue for review.
      parameters: [{ $ref: '#/components/parameters/ProductId' }, { $ref: '#/components/parameters/TenantHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PriceUpdateRequest' }
      responses:
        '200': { description: Applied or queued, content: { application/json: { schema: { $ref: '#/components/schemas/PriceUpdateResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Pending change exists, or product is not single-active-SKU }
  /catalog/products/{productId}/image:
    put:
      tags: [catalog]
      summary: Set/replace product image from a presigned-upload storageKey.
      parameters: [{ $ref: '#/components/parameters/ProductId' }, { $ref: '#/components/parameters/TenantHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [storageKey]
              properties:
                storageKey: { type: string, maxLength: 512 }
      responses:
        '200': { description: Image set, content: { application/json: { schema: { $ref: '#/components/schemas/ProductResponse' } } } }
        '400': { description: storageKey invalid or not owned by tenant }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [catalog]
      summary: Clear product image.
      parameters: [{ $ref: '#/components/parameters/ProductId' }, { $ref: '#/components/parameters/TenantHeader' }]
      responses:
        '204': { description: Cleared }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /marketing/promotions:
    get:
      tags: [marketing]
      summary: List promotions (VENDOR_ADMIN own vendor; OPS/ADMIN all incl. tenant-wide).
      parameters:
        - { $ref: '#/components/parameters/TenantHeader' }
        - { name: status, in: query, schema: { type: string } }
        - { name: vendorId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Promotion' } } } } }
    post:
      tags: [marketing]
      summary: Create a promotion. tenant-wide (vendor_id null) is OPS/ADMIN only.
      parameters: [{ $ref: '#/components/parameters/TenantHeader' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PromotionUpsert' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Promotion' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /marketing/promotions/{promotionId}:
    patch:
      tags: [marketing]
      summary: Update a promotion.
      parameters: [{ $ref: '#/components/parameters/PromotionId' }, { $ref: '#/components/parameters/TenantHeader' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PromotionUpsert' } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Promotion' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [marketing]
      summary: Delete a DRAFT/PAUSED promotion.
      parameters: [{ $ref: '#/components/parameters/PromotionId' }, { $ref: '#/components/parameters/TenantHeader' }]
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Cannot delete an ACTIVE promotion }
  /marketing/promotions/{promotionId}/activate:
    post:
      tags: [marketing]
      parameters: [{ $ref: '#/components/parameters/PromotionId' }, { $ref: '#/components/parameters/TenantHeader' }]
      responses: { '200': { description: Activated }, '409': { description: Expired or invalid } }
  /marketing/promotions/{promotionId}/pause:
    post:
      tags: [marketing]
      parameters: [{ $ref: '#/components/parameters/PromotionId' }, { $ref: '#/components/parameters/TenantHeader' }]
      responses: { '200': { description: Paused } }
...
  /reports/sales/summary:
    get:
      tags: [reports]
      parameters: [{ $ref: '#/components/parameters/TenantHeader' }, { $ref: '#/components/parameters/From' }, { $ref: '#/components/parameters/To' }, { $ref: '#/components/parameters/VendorIdQuery' }]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SalesSummary' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
  /reports/sales/timeseries:
    get:
      tags: [reports]
      parameters:
        - { $ref: '#/components/parameters/TenantHeader' }
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - { $ref: '#/components/parameters/VendorIdQuery' }
        - { name: interval, in: query, schema: { type: string, enum: [day, week, month], default: day } }
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/SalesBucket' } } } } } }
  /reports/sales/by-vendor:
    get:
      tags: [reports]
      parameters: [{ $ref: '#/components/parameters/TenantHeader' }, { $ref: '#/components/parameters/From' }, { $ref: '#/components/parameters/To' }]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/DimensionRow' } } } } } }
  /reports/sales/by-product:
    get:
      tags: [reports]
      parameters: [{ $ref: '#/components/parameters/TenantHeader' }, { $ref: '#/components/parameters/From' }, { $ref: '#/components/parameters/To' }, { $ref: '#/components/parameters/VendorIdQuery' }]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ProductDimensionRow' } } } } } }
  /reports/sales/by-category:
    get:
      tags: [reports]
      parameters: [{ $ref: '#/components/parameters/TenantHeader' }, { $ref: '#/components/parameters/From' }, { $ref: '#/components/parameters/To' }, { $ref: '#/components/parameters/VendorIdQuery' }]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/DimensionRow' } } } } } }
  /reports/sales/export:
    get:
      tags: [reports]
      summary: Stream a CSV of any report.
      parameters:
        - { $ref: '#/components/parameters/TenantHeader' }
        - { name: report, in: query, required: true, schema: { type: string, enum: [summary, timeseries, by-vendor, by-product, by-category] } }
        - { name: format, in: query, schema: { type: string, enum: [csv], default: csv } }
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - { $ref: '#/components/parameters/VendorIdQuery' }
      responses:
        '200': { description: CSV stream, content: { text/csv: { schema: { type: string, format: binary } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    TenantHeader: { name: X-Tenant-Id, in: header, required: false, schema: { type: string, format: uuid } }
...
    ProductId: { name: productId, in: path, required: true, schema: { type: string, format: uuid } }
...
    PromotionId: { name: promotionId, in: path, required: true, schema: { type: string, format: uuid } }
    From: { name: from, in: query, required: true, schema: { type: string, format: date-time } }
    To: { name: to, in: query, required: true, schema: { type: string, format: date-time } }
    VendorIdQuery: { name: vendorId, in: query, required: false, schema: { type: string, format: uuid } }
  responses:
    BadRequest: { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ProblemDetail' } } } }
    Forbidden: { description: Insufficient role or cross-tenant/vendor access }
    NotFound: { description: Not found in tenant scope }
  schemas:
    Money: { type: number, format: double, description: MZN, 2 decimals }
    ProblemDetail:
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors: { type: object, additionalProperties: { type: string } }
    ProductUpdateRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 140 }
        description: { type: string, maxLength: 2000 }
        categoryId: { type: string, format: uuid }
        requiresPrescription: { type: boolean }
    ProductResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string }
        requiresPrescription: { type: boolean }
        imageUrl: { type: string, nullable: true }
        price: { $ref: '#/components/schemas/Money' }
        pendingPrice: { allOf: [ { $ref: '#/components/schemas/Money' } ], nullable: true }
        updatedAt: { type: string, format: date-time }
    PriceUpdateRequest:
      type: object
      required: [price]
      properties: { price: { $ref: '#/components/schemas/Money' } }
    PriceUpdateResponse:
      type: object
      properties:
        skuId: { type: string, format: uuid }
        currentPrice: { $ref: '#/components/schemas/Money' }
        pendingPrice: { allOf: [ { $ref: '#/components/schemas/Money' } ], nullable: true }
        reviewRequired: { type: boolean }
...
    PromotionUpsert:
      type: object
      required: [name, type, value, scope, startsAt, endsAt]
      properties:
        vendorId: { type: string, format: uuid, nullable: true }
        name: { type: string, maxLength: 140 }
        code: { type: string, maxLength: 40, nullable: true }
        type: { type: string, enum: [PERCENTAGE, FIXED_AMOUNT] }
        value: { type: number, format: double }
        scope: { type: string, enum: [ORDER, CATEGORY, PRODUCT] }
        targetCategoryId: { type: string, format: uuid, nullable: true }
        targetProductId: { type: string, format: uuid, nullable: true }
        minOrderTotal: { allOf: [ { $ref: '#/components/schemas/Money' } ], nullable: true }
        maxDiscountAmount: { allOf: [ { $ref: '#/components/schemas/Money' } ], nullable: true }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        usageLimit: { type: integer, nullable: true }
        perCustomerLimit: { type: integer, nullable: true }
    Promotion:
      allOf:
        - $ref: '#/components/schemas/PromotionUpsert'
        - type: object
          properties:
            id: { type: string, format: uuid }
            usedCount: { type: integer }
            status: { type: string, enum: [DRAFT, ACTIVE, PAUSED, EXPIRED] }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    SalesSummary:
      type: object
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        orderCount: { type: integer, format: int64 }
        gross: { $ref: '#/components/schemas/Money' }
        discountTotal: { $ref: '#/components/schemas/Money' }
        refunds: { $ref: '#/components/schemas/Money' }
        net: { $ref: '#/components/schemas/Money' }
        commission: { $ref: '#/components/schemas/Money' }
        averageOrderValue: { $ref: '#/components/schemas/Money' }
        deliveredCount: { type: integer, format: int64 }
        cancelledCount: { type: integer, format: int64 }
    SalesBucket:
      allOf:
        - $ref: '#/components/schemas/SalesSummary'
        - type: object
          properties: { bucket: { type: string, format: date-time } }
    DimensionRow:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        gross: { $ref: '#/components/schemas/Money' }
        refunds: { $ref: '#/components/schemas/Money' }
        net: { $ref: '#/components/schemas/Money' }
        commission: { $ref: '#/components/schemas/Money' }
        sharePercent: { type: number, format: double }
    ProductDimensionRow:
      allOf:
        - $ref: '#/components/schemas/DimensionRow'
        - type: object
          properties: { quantitySold: { type: integer, format: int64 } }